+++ CHANGELOG.md
@@ -6,6 +6,7 @@ To install or update LODA, please follow the [installation instructions](https:/
 
 * Improve performance of `bin` operation by introducing a cache
 * Reduce maximum number of terms in validation to 4000 to increase performace
+* Omit expensive comparison for new program to increase performance
 
 ## v22.4.3